context を gin で、int64→int

+++ internal/domain/model/room.go
@@ -2,8 +2,8 @@ package model
 
 // Room は部屋
 type Room struct {
-	ID          int64
-	MaxUserNum  int64
+	ID          int
+	MaxUserNum  int
 	ChannelName string
 	// ThemeIds     []Theme
 }
